(wip) refactors users controller
